+++ src/assets/icons.js
@@ -57,4 +57,9 @@ const LockOpen = (props) => (
     <path d="M12 17c1.1 0 2-.9 2-2s-.9-2-2-2-2 .9-2 2 .9 2 2 2zm6-9h-1V6c0-2.76-2.24-5-5-5S7 3.24 7 6h1.9c0-1.71 1.39-3.1 3.1-3.1 1.71 0 3.1 1.39 3.1 3.1v2H6c-1.1 0-2 .9-2 2v10c0 1.1.9 2 2 2h12c1.1 0 2-.9 2-2V10c0-1.1-.9-2-2-2zm0 12H6V10h12v10z"/>
   </SvgIcon>
 );
-export { Bookmark, Wallpaper, Download, Upload, Extension, Settings, Link, About, History, Lock, LockOpen }
+const Permissions = (props) => (
+  <SvgIcon {...props}>
+    <path d="M12 1L3 5v6c0 5.55 3.84 10.74 9 12 5.16-1.26 9-6.45 9-12V5l-9-4zm-2 16l-4-4 1.41-1.41L10 14.17l6.59-6.59L18 9l-8 8z"/>
+  </SvgIcon>
+);
+export { Bookmark, Wallpaper, Download, Upload, Extension, Settings, Link, About, History, Lock, LockOpen, Permissions }