+++ src/pages-thumbnails.js
@@ -1,9 +1,0 @@
-function createThumbnail() {
-  chrome.tabs.onCreated.addListener(() => {
-    chrome.tabs.query({ highlighted: true }, (tabs) => {
-      chrome.tabs.captureVisibleTab({}, img => console.log(img))
-    })
-  })
-}
-
-createThumbnail();