+++ src/thumbnails.js
@@ -1,0 +1,13 @@
+"use strict";
+
+function test() {
+  chrome.runtime.sendMessage({ getThumbs: "" }, (thumbs) => {
+    this.setState({
+      thumbs: thumbs
+    })
+  });
+}
+
+module.exports = {
+  test: test
+}